+++ packages/yourloops/pages/signup/signup-formstate-context.tsx
@@ -28,10 +28,10 @@
 import React from "react";
 import _ from "lodash";
 
+import { HcpProfession } from "../../models/hcp-profession";
 import { UserRoles } from "../../models/shoreline";
 import { getCurrentLang } from "../../lib/language";
 import { SignupUser as IFormValues } from "../../lib/auth";
-import { HcpProfession } from "../../models/hcp-profession";
 
 export type FormValuesType = keyof IFormValues;
 
@@ -61,9 +61,9 @@ export const initialState: SignUpFormState = {
   formValues: {
     accountUsername: "",
     accountPassword: "",
-    accountRole: UserRoles.caregiver,
     profileFirstname: "",
     profileLastname: "",
+    accountRole: UserRoles.unset,
     profileCountry: "", // how to do better ?
     profilePhone: "",
     hcpProfession: HcpProfession.empty,